@@ -2,10 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use App\Menu;
+use App\Models\Menu;
 use Illuminate\Http\Request;
-use App\Http\Resources\MenuResource;
-use Illuminate\Support\Facades\Auth;
 
 class MenuController extends Controller
 {
@@ -16,16 +14,7 @@ class MenuController extends Controller
      */
     public function index()
     {
-        $menu = Menu::orderBy('sort_order')->get();
-        // return $menu;
-        return MenuResource::collection($menu);
-    }
-
-    public function list()
-    {
-        $menu = Menu::orderBy('sort_order')->get();
-        $data['menus'] = $menu;
-        return view('listMenu')->with($data);
+        //
     }
 
     /**
@@ -46,49 +35,44 @@ class MenuController extends Controller
      */
     public function store(Request $request)
     {
-        if (!Auth::check()) {
-            return 'unauthorized';
-        }
         $menu_data = [
             'en' => ['name'  => $request->en_name],
             'es' => ['name' => $request->es_name,],
             'sort_order' => $request->sort_order
         ];
         $m = Menu::create($menu_data);
-        if ($request->web == true) {
-            return redirect(route('admin.menulist'));
-        }
-        return new MenuResource($m);
+
+        return redirect(route('admin'));
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Menu  $menu
+     * @param  \App\Models\Menu  $menu
      * @return \Illuminate\Http\Response
      */
     public function show(Menu $menu)
     {
-        return new MenuResource($menu);
+        //
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Menu  $menu
+     * @param  \App\Models\Menu  $menu
      * @return \Illuminate\Http\Response
      */
     public function edit(Menu $menu)
     {
         $data['menu'] = $menu;
-        return view('newadmin.menuedit')->with($data);
+        return view('admin.menu.edit')->with($data);
     }
 
     /**
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Menu  $menu
+     * @param  \App\Models\Menu  $menu
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Menu $menu)
@@ -98,25 +82,20 @@ class MenuController extends Controller
             'es' => ['name' => $request->es_name,],
             'sort_order' => $request->sort_order
         ];
+
         $menu->update($menu_data);
-        if ($request->web == true) {
-            return redirect(route('admin.menuedit', $menu->id));
-        }
-        return new MenuResource($menu);
+        return redirect((route('admin')));
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Menu  $menu
+     * @param  \App\Models\Menu  $menu
      * @return \Illuminate\Http\Response
      */
     public function destroy(Menu $menu)
     {
         $menu->delete();
-        if (request('web') == true) {
-            return redirect(route('admin.menulist'));
-        }
-        return $menu;
+        return redirect(route('admin'));
     }
 }
