@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\OrderItem;
+use App\Models\OrderItem;
 use Illuminate\Http\Request;
 
 class OrderItemController extends Controller
@@ -41,7 +41,7 @@ class OrderItemController extends Controller
     /**
      * Display the specified resource.
      *
-     * @param  \App\OrderItem  $orderItem
+     * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
     public function show(OrderItem $orderItem)
@@ -52,7 +52,7 @@ class OrderItemController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\OrderItem  $orderItem
+     * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
     public function edit(OrderItem $orderItem)
@@ -64,7 +64,7 @@ class OrderItemController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\OrderItem  $orderItem
+     * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, OrderItem $orderItem)
@@ -75,7 +75,7 @@ class OrderItemController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\OrderItem  $orderItem
+     * @param  \App\Models\OrderItem  $orderItem
      * @return \Illuminate\Http\Response
      */
     public function destroy(OrderItem $orderItem)
