@@ -2,7 +2,6 @@
 
 namespace App\Http\Requests;
 
-use App\Rules\deliveryTime;
 use Illuminate\Foundation\Http\FormRequest;
 
 class RegisterUnit extends FormRequest
@@ -20,7 +19,7 @@ class RegisterUnit extends FormRequest
     /**
      * Get the validation rules that apply to the request.
      *
-     * @return array
+     * @return array<string, mixed>
      */
     public function rules()
     {
@@ -30,7 +29,6 @@ class RegisterUnit extends FormRequest
                 'in:101, 102, 103, 104, 201, 202, 203, 301, 302, 303, 401, 402, 403, 501, 502, 601, 602, 701, 702, 703, 801, 802, 803, 901, 902, 903, 1001, 1002, 1003, 1101, 1102, 1201'
             ],
             'name' => ['required'],
-            'time' => ['required', new deliveryTime()],
         ];
     }
 
