@@ -1,133 +1,57 @@
-@extends('layouts.app')
-
-
-@section('content')
-
-<div class="container">
-
-    <div class="row justify-content-center">
-
-        <div class="col-md-8">
-
-            <div class="card">
-
-                <div class="card-header">{{ __('Login') }}</div>
-
-
-                <div class="card-body">
-
-                    <form method="POST" action="{{ route('login') }}">
-
-                        @csrf
-
-
-                        <div class="form-group row">
-
-                            <label for="email" class="col-md-4 col-form-label text-md-right">{{ __('E-Mail Address') }}</label>
-
-
-                            <div class="col-md-6">
-
-                                <input id="email" type="email" class="form-control @error('email') is-invalid @enderror" name="email" value="{{ old('email') }}" required autocomplete="email" autofocus>
-
-
-                                @error('email')
-
-                                    <span class="invalid-feedback" role="alert">
-
-                                        <strong>{{ $message }}</strong>
-
-                                    </span>
-
-                                @enderror
-
-                            </div>
-
-                        </div>
-
-
-                        <div class="form-group row">
-
-                            <label for="password" class="col-md-4 col-form-label text-md-right">{{ __('Password') }}</label>
-
-
-                            <div class="col-md-6">
-
-                                <input id="password" type="password" class="form-control @error('password') is-invalid @enderror" name="password" required autocomplete="current-password">
-
-
-                                @error('password')
-
-                                    <span class="invalid-feedback" role="alert">
-
-                                        <strong>{{ $message }}</strong>
-
-                                    </span>
-
-                                @enderror
-
-                            </div>
-
-                        </div>
-
-
-                        <div class="form-group row">
-
-                            <div class="col-md-6 offset-md-4">
-
-                                <div class="form-check">
-
-                                    <input class="form-check-input" type="checkbox" name="remember" id="remember" {{ old('remember') ? 'checked' : '' }}>
-
-
-                                    <label class="form-check-label" for="remember">
-
-                                        {{ __('Remember Me') }}
-
-                                    </label>
-
-                                </div>
-
-                            </div>
-
-                        </div>
-
-
-                        <div class="form-group row mb-0">
-
-                            <div class="col-md-8 offset-md-4">
-
-                                <button type="submit" class="btn btn-primary">
-
-                                    {{ __('Login') }}
-
-                                </button>
-
-
-                                @if (Route::has('password.request'))
-
-                                    <a class="btn btn-link" href="{{ route('password.request') }}">
-
-                                        {{ __('Forgot Your Password?') }}
-
-                                    </a>
-
-                                @endif
-
-                            </div>
-
-                        </div>
-
-                    </form>
-
-                </div>
-
+<x-guest-layout>
+    <x-auth-card>
+        <x-slot name="logo">
+            <a href="/">
+                <x-application-logo class="w-20 h-20 fill-current text-gray-500" />
+            </a>
+        </x-slot>
+
+        <!-- Session Status -->
+        <x-auth-session-status class="mb-4" :status="session('status')" />
+
+        <form method="POST" action="{{ route('login') }}">
+            @csrf
+
+            <!-- Email Address -->
+            <div>
+                <x-input-label for="email" :value="__('Email')" />
+
+                <x-text-input id="email" class="block mt-1 w-full" type="email" name="email" :value="old('email')" required autofocus />
+
+                <x-input-error :messages="$errors->get('email')" class="mt-2" />
             </div>
 
-        </div>
+            <!-- Password -->
+            <div class="mt-4">
+                <x-input-label for="password" :value="__('Password')" />
 
-    </div>
+                <x-text-input id="password" class="block mt-1 w-full"
+                                type="password"
+                                name="password"
+                                required autocomplete="current-password" />
 
-</div>
+                <x-input-error :messages="$errors->get('password')" class="mt-2" />
+            </div>
 
-@endsection
+            <!-- Remember Me -->
+            <div class="block mt-4">
+                <label for="remember_me" class="inline-flex items-center">
+                    <input id="remember_me" type="checkbox" class="rounded border-gray-300 text-indigo-600 shadow-sm focus:border-indigo-300 focus:ring focus:ring-indigo-200 focus:ring-opacity-50" name="remember">
+                    <span class="ml-2 text-sm text-gray-600">{{ __('Remember me') }}</span>
+                </label>
+            </div>
+
+            <div class="flex items-center justify-end mt-4">
+                @if (Route::has('password.request'))
+                    <a class="underline text-sm text-gray-600 hover:text-gray-900" href="{{ route('password.request') }}">
+                        {{ __('Forgot your password?') }}
+                    </a>
+                @endif
+
+                <x-primary-button class="ml-3">
+                    {{ __('Log in') }}
+                </x-primary-button>
+            </div>
+        </form>
+    </x-auth-card>
+</x-guest-layout>
