@@ -1,11 +1,11 @@
 @extends('layouts.app')
 
 @section('content')
-    
-    <div class="container">
-            <div class="card">
-                <div class="card-header">{{ __('Login') }}    
-                    <h4>
+
+    <div class="w-full">
+            <div class="card mx-auto">
+                <div class="card-header">{{ __('Login') }}
+                    <h4 class="font-semibold text-xl">
                         {{__('messages.order',['startTime' => $openTime, 'endTime'=>$closeTime])}}
                     </h4>
                 </div>
@@ -13,12 +13,12 @@
                 <div class="card-body">
                     <form method="POST" action="{{ route('register') }}">
                         @csrf
-                        <div class="form-group">
+                        <div class="form-group text-center">
                             <label for="unit">{{__('Unit Number')}}:</label>
-                            <input type="input" name="unit" value="{{old('unit')}}">
+                            <input type="text" name="unit" value="{{old('unit')}}">
                             <label for="name">{{__('Last Name')}}:</label>
-                            <input type="input" name="name" value="{{old('name')}}">
-                            <label for="time">{{__('Requested Delivery Time')}} 
+                            <input type="text" name="name" value="{{old('name')}}">
+                            <label for="time">{{__('Requested Delivery Time')}}
                                 {{__('messages.delivery', [
                                     'startTime' => $deliveryStart,  'endTime'=>$deliveryEnd])}}<br> ({{__('Please allow at least 30 minutes')}}): </label>
                                     <select name="time" id="time">
@@ -28,18 +28,18 @@
                                     </select>
 
                             <div class="form-check">
-                            <label class="form-check-label"> 
+                            <label class="form-check-label">
                                 <input class="form-check-input"  type="radio" name="location" value="Palapa" id="location1" {{ (old('location') == 'Palapa' || old('location')=='') ? 'checked' : '' }}>
                                     {{__('I will eat at the Palapa')}} &mdash; {{__('use glass dishes')}}
                                 </label>
-                                
+
                                 <label class="form-check-label">
-                                    <input class="form-check-input" type="radio" name="location" value="Para llevar" id="location2" {{ (old('location') == 'Para llevar') ? 'checked' : '' }}>  
+                                    <input class="form-check-input" type="radio" name="location" value="Para llevar" id="location2" {{ (old('location') == 'Para llevar') ? 'checked' : '' }}>
 		                            {{__('I will pick up from the Palapa')}} &mdash; {{__('use paper plates')}}
-                                </label>   
-                            </div>                                 
+                                </label>
+                            </div>
                         </div>
-                        <div class="form-group">
+                        <div class="form-group text-center">
                                 <button type="submit" class="btn btn-primary">
                                     {{ __('Login') }}
                                 </button>
@@ -54,7 +54,7 @@
             @endforeach
         </ul>
     </div>
-@endif                    
+@endif
                 </div>
             </div>
         </div>
