@@ -7,7 +7,33 @@
             <h4>{{__('The Palapa is now closed for on-line orders')}}</h4>
         </div>
     </div>
-    @include('menu')
+
+    @if ($special)
+    <aside id="popUp" class="popup">
+        <div class="popUpContainer">
+            <header>
+                <a href="#!" class="closePopUp">X</a>
+                <h2>{{$special_title}}</h2>
+            </header>
+            <article>
+                <p>{{$special_message}}</p>
+            </article>
+        </div>
+        <a href="#!" class="closePopUpOutSide"></a>
+    </aside>
+@endif
+    @include('partials.menu')
 </div>
+
+<script>
+    const closeModal = document.querySelector('.closePopUp');
+    const modal = document.querySelector('#popUp');
+
+    if (modal) {
+            closeModal.addEventListener('click', ()=> {
+            modal.style.display="none";
+        });
+    }
+</script>
 
 @endsection
