@@ -14,13 +14,6 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:api')->get('/user', function (Request $request) {
+Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-
-Route::resource('menu', 'MenuController');
-Route::resource('category', 'CategoryController');
-Route::resource('item', 'ItemController');
-Route::resource('option', 'OptionController');
-Route::resource('choice', 'ChoiceController');
-Route::resource('selection', 'SelectionController');
